Ikke forsøk å slette vedlegg hvis det ikke er noen vedlegg å slette i mellomlagring

+++ apps/foreldrepengesoknad/src/app/steps/oppsummering/Oppsummering.tsx
@@ -139,7 +139,9 @@ const Oppsummering = () => {
 
                 try {
                     await Api.deleteMellomlagretSøknad(fnr, abortSignal);
-                    await Api.deleteMellomlagredeVedlegg(fnr, cleanedSøknad.vedlegg, abortSignal);
+                    if (cleanedSøknad.vedlegg.length > 0) {
+                        await Api.deleteMellomlagredeVedlegg(fnr, cleanedSøknad.vedlegg, abortSignal);
+                    }
                 } catch (error) {
                     // Vi bryr oss ikke om feil her. Logges bare i backend
                 }
